Migrate subscribtions to TS (WIP)
