--- src/openapi/streaming/subscription-actions.js
@@ -1,5 +1,0 @@
-export const ACTION_SUBSCRIBE = 0x1;
-export const ACTION_UNSUBSCRIBE = 0x2;
-export const ACTION_MODIFY = 0x4;
-export const ACTION_MODIFY_PATCH = 0x8;
-export const ACTION_UNSUBSCRIBE_BY_TAG_PENDING = 0x10;
